# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TrackFlow API — digital tracking governance platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces/{workspaceId}/campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all campaigns in a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [active, paused, completed, draft]
        - name: platform
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a new campaign
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /workspaces/{workspaceId}/campaigns/{campaignId}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign by ID
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /workspaces/{workspaceId}/campaigns/{campaignId}/stats:
    get:
      operationId: getCampaignStats
      tags: [campaigns]
      summary: Get performance stats for a campaign
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignStats"

  # ─── Links ────────────────────────────────────────────────────────────────
...
  /workspaces/{workspaceId}/conversions:
    get:
      operationId: listConversions
      tags: [conversions]
      summary: List all conversions
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: campaignId
          in: query
          schema:
            type: ["integer", "null"]
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of conversions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversion"
    post:
      operationId: createConversion
      tags: [conversions]
      summary: Record a new conversion
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversion"
  /workspaces/{workspaceId}/conversions/summary:
    get:
      operationId: getConversionSummary
      tags: [conversions]
      summary: Get conversion summary grouped by campaign/platform/creative
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Conversion summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversionSummary"
components:
  schemas:
...
    Campaign:
      type: object
      required: [id, workspaceId, name, status, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        name:
          type: string
        client:
          type: ["string", "null"]
        platform:
          type: ["string", "null"]
        objective:
          type: ["string", "null"]
        responsible:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, paused, completed, draft]
        budget:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    CampaignInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        client:
          type: string
        platform:
          type: string
        objective:
          type: string
        responsible:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, paused, completed, draft]
        budget:
          type: number
        notes:
          type: string
        startDate:
          type: string
        endDate:
          type: string
    CampaignUpdate:
      type: object
      properties:
        name:
          type: string
        client:
          type: string
        platform:
          type: string
        objective:
          type: string
        responsible:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, paused, completed, draft]
        budget:
          type: number
        notes:
          type: string
        startDate:
          type: string
        endDate:
          type: string
    CampaignStats:
      type: object
      required: [campaignId, campaignName, clicks, leads, conversions, revenue, roas]
      properties:
        campaignId:
          type: integer
        campaignName:
          type: string
        clicks:
          type: integer
        leads:
          type: integer
        conversions:
          type: integer
        revenue:
          type: number
        roas:
          type: number

    # ─── Links ─────────────────────────────────────────────────────────────
...
    Conversion:
      type: object
      required: [id, workspaceId, revenue, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        campaignId:
          type: ["integer", "null"]
        campaignName:
          type: ["string", "null"]
        leadId:
          type: ["string", "null"]
        revenue:
          type: number
        platform:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        medium:
          type: ["string", "null"]
        creative:
          type: ["string", "null"]
        adSpend:
          type: ["number", "null"]
        createdAt:
          type: string
    ConversionInput:
      type: object
      required: [revenue]
      properties:
        campaignId:
          type: integer
        leadId:
          type: string
        revenue:
          type: number
        platform:
          type: string
        source:
          type: string
        medium:
          type: string
        creative:
          type: string
        adSpend:
          type: number
    ConversionSummary:
      type: object
      required: [totalRevenue, totalConversions, averageTicket, byCampaign, byPlatform]
      properties:
        totalRevenue:
          type: number
        totalConversions:
          type: integer
        averageTicket:
          type: number
        byCampaign:
          type: array
          items:
            $ref: "#/components/schemas/ConversionBucket"
        byPlatform:
          type: array
          items:
            $ref: "#/components/schemas/ConversionBucket"
        byCreative:
          type: array
          items:
            $ref: "#/components/schemas/ConversionBucket"
    ConversionBucket:
      type: object
      required: [label, conversions, revenue]
      properties:
        label:
          type: string
        conversions:
          type: integer
        revenue:
          type: number
        roas:
          type: ["number", "null"]