# lib/api-zod/src/generated/types/conversionInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * TrackFlow API — digital tracking governance platform
 * OpenAPI spec version: 0.1.0
 */

export interface ConversionInput {
  campaignId?: number;
  leadId?: string;
  revenue: number;
  platform?: string;
  source?: string;
  medium?: string;
  creative?: string;
  adSpend?: number;
}